services/data-fusion-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
xarray==2023.11.0
netcdf4==1.6.5
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
httpx==0.25.2
